refactor: Controller 추출
