+++ app/src/main/java/nextstep/jwp/constants/Http.java
@@ -2,7 +2,8 @@ package nextstep.jwp.constants;
 
 public class Http {
     public static final String SEPARATOR = " ";
-    public static final String EMPTY_LINE =  "\r\n";
+    public static final String NEW_LINE =  "\r\n";
+    public static final String EMPTY_LINE = "";
     public static final String FILE_EXTENSION_SEPARATOR = "\\.";
     public static final String DOT = ".";
     public static final String FILE_EXTENSION_HTML = ".html";
