+++ app/src/main/java/nextstep/jwp/constants/HttpMethod.java
@@ -1,0 +1,6 @@
+package nextstep.jwp.constants;
+
+public class HttpMethod {
+    public static final String GET = "GET";
+    public static final String POST = "POST";
+}
